corrección del rawkernel. La dirección vertical estaba invertida, ahora el código es exactamente equivalente al de elementwise

+++ fuego_referencia.py
@@ -1,5 +1,4 @@
-from math import sqrt
-from modelo_rdc import spread_infection, spread_infection_raw
+from modelo_rdc import spread_infection_raw
 import numpy as np 
 import cupy as cp # type: ignore
 
@@ -40,35 +39,37 @@ ny, nx = vegetacion.shape  # Usamos cualquier mapa para obtener las dimensiones
 ############################## PARÁMETROS DEL INCENDIO DE REFERENCIA ###############################################
 
 # Tamaño de cada celda
-d = 30 # metros
+d = cp.float32(30) # metros
 # Coeficiente de difusión
-D = 50 # metros^2 / hora. Si la celda tiene 30 metros, en una hora avanza 1/3 del tamaño de la celda
+D = cp.float32(50) # metros^2 / hora. Si la celda tiene 30 metros, en una hora avanza 1/3 del tamaño de la celda
 
 # Parámetros del modelo SI
 beta_veg = cp.where(vegetacion <= 2, 0, 0.1 * vegetacion) # fracción de vegetación incéndiandose por hora
 
 # Hacemos una máscara donde vegetación <=2, gamma >> 1/dt. Sino, vale 0.1. 
-gamma = cp.where(vegetacion <= 2, cp.float32(100), cp.float32(0.1))
-#gamma = cp.where(vegetacion <= 2, 100, 0.1) # fracción de vegetación incéndiandose por hora.
-            # 1/gamma es el tiempo promedio del incendio
+gamma = cp.where(vegetacion <= 2, cp.float32(100), cp.float32(0.1)) # fracción de vegetación incéndiandose por hora.
+                                                                    # 1/gamma es el tiempo promedio del incendio
 
-dt = 1/6 # Paso temporal. Si medimos el tiempo en horas, 1/6 indica un paso de 10 minutos
+beta_veg = beta_veg.astype(cp.float32)
+gamma = gamma.astype(cp.float32)
+
+dt = cp.float32(1/6) # Paso temporal. Si medimos el tiempo en horas, 1/6 indica un paso de 10 minutos
 
 # Transformación del viento a coordenadas cartesianas
 # El viento está medido en km/h. En m/h el viento es una cantidad enorme, por eso
 # la cantidad A que multiplica el viento tendría que ser pequeña. Intuición: A~10^-5
-wx = vientov * cp.cos(5/2 * cp.pi - vientod * cp.pi / 180) * 1000
-wy = - vientov * cp.sin(5/2 * cp.pi - vientod * cp.pi / 180) * 1000
+wx = (vientov * cp.cos(5/2 * cp.pi - vientod * cp.pi / 180) * 1000).astype(cp.float32)
+wy = (- vientov * cp.sin(5/2 * cp.pi - vientod * cp.pi / 180) * 1000).astype(cp.float32)
 
 # Constante A adimensional
-A = 5e-4 # 10^-3 está al doble del límite de estabilidad
+A = cp.float32(5e-4) # 10^-3 está al doble del límite de estabilidad
 
 # Constante B
-B = 15 # m/h
+B = cp.float32(15) # m/h
 
 # Cálculo de la pendiente (usando mapas de pendiente y orientación)
-h_dx_mapa = cp.tan(pendiente * cp.pi / 180) * cp.cos(orientacion * cp.pi / 180 - cp.pi/2)
-h_dy_mapa = cp.tan(pendiente * cp.pi / 180) * cp.sin(orientacion * cp.pi / 180 - cp.pi/2)
+h_dx_mapa = (cp.tan(pendiente * cp.pi / 180) * cp.cos(orientacion * cp.pi / 180 - cp.pi/2)).astype(cp.float32)
+h_dy_mapa = (cp.tan(pendiente * cp.pi / 180) * cp.sin(orientacion * cp.pi / 180 - cp.pi/2)).astype(cp.float32)
 
 ############################## INCENDIO DE REFERENCIA ###############################################
 
@@ -85,10 +86,10 @@ var_poblacion = 0
 
 # Inicializar arrays de cupy para almacenar los resultados
 num_steps = 1001
-#pob_total = cp.zeros(num_steps)
-#S_total = cp.zeros(num_steps)
-#I_total = cp.zeros(num_steps)
-#R_total = cp.zeros(num_steps)
+# pob_total = cp.zeros(num_steps)
+# S_total = cp.zeros(num_steps)
+# I_total = cp.zeros(num_steps)
+# R_total = cp.zeros(num_steps)
 
 start = cp.cuda.Event()
 end = cp.cuda.Event()
@@ -109,24 +110,24 @@ for t in range(num_steps):
     I, I_new = I_new, I
     R, R_new = R_new, R
 
-    #suma_S = S.sum() / nx**2
-    #suma_I = I.sum() / nx**2
-    #suma_R = R.sum() / nx**2
+    # suma_S = S.sum() / (nx*ny)
+    # suma_I = I.sum() / (nx*ny)
+    # suma_R = R.sum() / (nx*ny)
 
-    #suma_total = suma_S + suma_I + suma_R
-    #pob_total[t] = suma_total
-    #S_total[t] = suma_S
-    #I_total[t] = suma_I
-    #R_total[t] = suma_R
+    # suma_total = suma_S + suma_I + suma_R
+    # pob_total[t] = suma_total
+    # S_total[t] = suma_S
+    # I_total[t] = suma_I
+    # R_total[t] = suma_R
 
-    #var_poblacion += cp.abs(suma_total - pob_total[t-1]) if t > 0 else 0
+    # var_poblacion += cp.abs(suma_total - pob_total[t-1]) if t > 0 else 0
 
 end.record()  # Marca el final en GPU
 end.synchronize() # Sincroniza y mide el tiempo
 
-#var_poblacion_promedio = var_poblacion / num_steps
+# var_poblacion_promedio = var_poblacion / num_steps
 
-#print(f'Variación de población promedio: {var_poblacion_promedio}')
+# print(f'Variación de población promedio: {var_poblacion_promedio}')
 
 # Calcular el número de celdas quemadas
 celdas_quemadas = cp.sum(R > 0.2)
